requirements.txt
opencv-python==4.8.1
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.3
matplotlib==3.8.2